detach from su on rt_sigprocmask(SIG_SETMASK, ...)
The previous code that detaches on clone would miss the password on old
CentOS systems. Instead we detach on rt_sigprocmask(SIG_SETMASK, ...)
which occurs immediately after the password is captured. It is is still
called if no password is required (as in the case of root calling su).

Along with the previous commits, fixes #2